Initial port of hi4socet, hinoproj from Perl to Python (#67)
Naive ports of the legacy hi4socet.pl and hinoproj.pl to Python. The ported code replicates the functionality and naming conventions of the original Perl with 2 minor exceptions:
 - Python code does not clean up (delete) text flatfiles generated by hijitreg or cubes generated by noproj after mosaicking
 - The order of ISIS programs called by hi4socet has been changed from the original (but this does not change output)

As in the original Perl, hinoproj.py is designed to be run as a standalone script or called from hi4socet. This is accomplished in an ugly, non-Pythonic way in the current commit in order to create a minimum viable port quickly.