SurfaceFit: Add GXP Support and Simplify Code (#73)
Updated `surface_fit_pc_align.py` and `gpf_transform.py` to accept an option `--gxp` to indicate that input GPF file is in Socet GXP format. Default behavior is still to treat input GPF as legacy Socet Set format. This closes #72
Moved functions that were duplicated in the above scripts into a `surfacefit` module.
Added above scripts to script list in `setup.py` so that they are added to `PATH` on installation.